MODIFIED: based on the outlined specifications.
Found out that init_disk needs to be private to blockio. Wipe_disk was created to take it's place. Also, the free_block list has now been completely been replaced by a boolean array. Also, some notes were added to the initialize.c to outline what needs to be done inorder to complete intialize the drive. Some of the methods need are currently not implmented (such as create()). Finally, superblock.h was modified to completed remove the freeblock struct.