Added the complete concat function, needs testing
Added the concat function, which is used to concatenate two arrays and return
a destination pointer containing the results of the two. It will automatically
realloc the appropriately sized array to store the results in.

Note, in order to make it a generic for more than just one type it takes void pointers, it's a prime example of when static type systems break down in generic
programming