MODIFIED: fixed a bug in unlink inode
Fixed a bug where the new array of data locations would contain a null space for the removed element. Fixed a bug where the fbl would be update before the inode was unlinked from the parent and blocks that where part of the deleted file would be over written thus breaking copy on write.