MODIFIED: fixed up bug with root delete seg fault
Major hack, at the start of delete the pathname is checked if the length of it is equal to 1. If so the only valid path that could be given is root, (since everything else given would be invalid because it does not reference root). Therefore a general error is through, invalid path.